Renames to no-op folders
