+++ src/unstract/sdk/adapters/embedding/no_op/src/__init__.py
