+++ src/unstract/sdk/adapters/embedding/no_op/src/no_op_embedding.py
