--- src/unstract/sdk/adapters/llm/no_op_llm/README.md
@@ -1,3 +1,0 @@
-# Unstract NoOp LLM adapter for load testing
-
-An LLM adapter that does not perform any operation. Waits for the configured time before returning a response. This can be useful to perform tests on the system
