+++ src/unstract/sdk/adapters/vectordb/no_op/README.md
@@ -1,0 +1,3 @@
+# Unstract NoOp vectordb adapter for load testing
+
+A vectodb adapter that does not perform any operation. Waits for the configured time before returning a response. This can be useful to perform tests on the system
