+++ src/unstract/sdk/adapters/vectordb/no_op/src/static/json_schema.json
@@ -1,5 +1,5 @@
 {
-  "title": "NoOp VectorDB Adapter",
+  "title": "No-op",
   "type": "object",
   "required": [
     "adapter_name",
