--- src/unstract/sdk/adapters/vectordb/no_op_vectordb/README.md
@@ -1,3 +1,0 @@
-# Unstract NoOp vectordb adapter for load testing
-
-A vectodb adapter that does not perform any operation. Waits for the configured time before returning a response. This can be useful to perform tests on the system
