+++ src/unstract/sdk/adapters/x2text/no_op/src/no_op_x2text_adapter.py
