+++ src/unstract/sdk/adapters/x2text/no_op/src/static/json_schema.json
@@ -1,5 +1,5 @@
 {
-  "title": "NoOp X2Text Convertor",
+  "title": "No-op",
   "type": "object",
   "required": [
     "adapter_name",
