--- src/unstract/sdk/adapters/x2text/no_op_x2text/README.md
@@ -1,3 +1,0 @@
-# Unstract NoOp x2text adapter for load testing
-
-A x2text adapter that does not perform any operation. Waits for the configured time before returning a response. This can be useful to perform tests on the system
